Refactored the filter of the Account Type selector
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C33865

+++ Retail/Code/Retail.BusinessEntity.Entities/AccountType/AccountTypeFilter.cs
@@ -8,6 +8,6 @@ namespace Retail.BusinessEntity.Entities
 {
     public class AccountTypeFilter
     {
-        public int? ParentAccountId { get; set; }
+        public long? ParentAccountId { get; set; }
     }
 }
